new: building ability scores with die rolls
The Pathfinder game specifies a few ways to randomly generate a set of
AbilityScores: the classical way, the modern way, and the heroic way.

The classical way, dating back to the earlier AD&D days, involve
rolling three six-sided dice (3d6) and summing the results.
The modern way, the way modern D&D does it as of the fifth edition and
Pathfinder recommends, is to roll 4d6 and not include the smallest roll.
This often gets you better scores than the classical way.
The heroic way is to simply roll 2d6 and add 6 to the result.
This often gets you better scores than the modern way.

To call, import the BuildMethod and use one to roll ability scores.

import network.cardboard.crystallogic.AbilityScores.BuildMethod;
// ...
AbilityScores stats = new AbilityScores(BuildMethod.MODERN);