fix: JSON parsing constructor, JSON parsing test
Ignoring the fact we need to write more tests, the fact that the JSON
structure keeps changing means that in some cases, the program won't
compile because the tests weren't updated to reflect it!
Methods and constructors can be overloaded by type.
Constructors have been separated into two: one having a String type
and one on a JsonNode type.